customer_service/pyproject.toml
[project]
name = "face-verification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "absl-py==2.3.1",
    "art>=6.5",
    "attrs==25.3.0",
    "cffi==1.17.1",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "et-xmlfile==2.0.0",
    "flatbuffers==25.2.10",
    "fonttools==4.58.5",
    "jax==0.6.2",
    "jaxlib==0.6.2",
    "kiwisolver==1.4.8",
    "matplotlib==3.10.3",
    "mediapipe==0.10.9",
    "ml-dtypes==0.5.1",
    "numpy==2.2.6",
    "opencv-contrib-python==4.12.0.88",
    "opencv-python==4.12.0.88",
    "openpyxl==3.1.5",
    "opt-einsum==3.4.0",
    "packaging==25.0",
    "pandas==2.3.1",
    "pika>=1.3.2",
    "pillow==11.3.0",
    "pip==25.1.1",
    "protobuf>=3.11,<4",
    "pycparser==2.22",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.1.1",
    "pytz==2025.2",
    "PyYAML==6.0.2",
    "rich>=13.8.0",
    "scipy==1.16.0",
    "setuptools==65.5.0",
    "six==1.17.0",
    "sounddevice==0.5.2",
    "tzdata==2025.2",
]
